move things to cogs instead
